--- lib/config/production.js
@@ -1,14 +1,0 @@
-"use strict";
-
-module.exports = {
-  "electrode": {
-    "source": "production"
-  },
-  "connections": {
-    "default": {
-      "routes": {
-        "cors": false
-      }
-    }
-  }
-};
