+++ src/config/production.ts
@@ -1,0 +1,12 @@
+export = {
+  electrode: {
+    source: "production"
+  },
+  connections: {
+    default: {
+      routes: {
+        cors: false
+      }
+    }
+  }
+};
