+++ tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "compilerOptions": {
+    "outDir": "lib",
+    "lib": [
+      "es6"
+    ],
+    "module": "CommonJS",
+    "esModuleInterop": true,
+    "target": "es6",
+    "preserveConstEnums": true,
+    "sourceMap": true,
+    "declaration": true,
+    "types": [
+      "node"
+    ],
+    "forceConsistentCasingInFileNames": true,
+    "noImplicitReturns": true,
+    "alwaysStrict": true,
+    "strictFunctionTypes": true
+  },
+  "include": [
+    "src"
+  ]
+}
