--- xclap.js
@@ -1,1 +1,0 @@
-require("electrode-archetype-njs-module-dev")();
