+++ xclap.ts
@@ -1,0 +1,2 @@
+import { loadTasks } from "@xarc/module-dev";
+loadTasks();
